이동을 캐릭터 속도에 맞춰서 이동

+++ YaMang/YaMangServer/YaMangServer/Corps.cpp
@@ -18,11 +18,11 @@ Corps::~Corps()
 {
 }
 
-void Corps::AddDamage( int damage )
+void Corps::AddDamage( float damage )
 {
 	if ( !m_IsDead )
 	{
-		int defense = m_Defense + m_DefenseBonus - damage;
+		float defense = m_Defense + m_DefenseBonus - damage;
 		if ( defense < 0 )
 		{
 			m_HP = m_HP + defense;
